Hardcode GitHub org in View on OpenSAFELY URL
We're seeing users find the substitutions a little tricky, and to make matters worse confuse GitHub orgs with OpenSAFELY orgs.  Since we currently have no repos hosted in external orgs we're going to remove one of the steps for researchers.

Note: I've left the setup.yaml workflow as is so we can re-add the template string later if we need to.