Add Python requirements
For now, switch to a list of requirements instead of pinning versions
and hashes.

That way, we get the latest on container creation. This is fine for
current development and testing.

Certainly, having the latest OpenSAFELY CLI version is probably what we
want in any case.

Having the latest ehrQL version is fine for now, but may not be what we
want in future. What if we move to ehrQL v2 and a user is still using
ehrQL v1?

It's not clear right now what the best way of managing
the ehrQL version would be for installing it as a Python package:

* user maintenance chore?
* automated workflow?
* specifying the ehrQL version via some other way, and passing that into
  a post-container build setup script?