Provide a custom dev container image
This image is based on Tom Palmer's rebuild of the r-docker image, used in the
OpenSAFELY pipeline. It provides the same version of R and copies over all
required packages. It also packages RStudio for use with devcontainers.

The image also provides the version of Python used by the OpenSAFELY pipeline
and copies over the required Python packages.

The ehrql library isn't compatible with this version of Python, so isn't
installed.

Co-authored-by: Jon Massey <jon.massey@thedatalab.org>