Disable git autofetch
Git autofetch allows commits to be automatically fetched from the default
remote. If it's not set, then a user will be prompted to set it whenever they
create a new Codespace. This prompt is annoying and means we need to document
what action a user should take. Therefore, we're setting a sensible default.
This change turns the autofetch off, as we believe users expect to fetch any
changes manually and it has the potential to cause confusion if it goes wrong.