Remove docker-in-docker options that are defaults
These four removed options are currently the default configuration for
the feature.

If we need to override the defaults in future, then we should do that.

Reference:

https://github.com/devcontainers/features/blob/f5787eed01022f177475a99084327e023a84ddaf/src/docker-in-docker/README.md

These explicit options were added in ef7549f48bee0debf63e97a87a93feb8e5ea1e0a
and that in turn was because Visual Studio Code added them in when
creating the dev container configuration.

For context,

* `moby` specifies whether use Moby instead of Docker CE
* `version` specifies the Moby/Docker CE version
* `azureDnsAutoDetection` sets the DNS server for `dockerd` if running
  in Azure (Actions runners are hosted in Azure, not sure if this
  setting makes any difference for our use)
* `installDockerBuildx`: installs Docker Buildx (which is used by Docker
  Buildkit); not convinced there's anywhere we use this for OpenSAFELY
  CLI, but it shouldn't do any harm to have it installed