Handle empty rstudio-prefs.json
Based on [feedback](https://bennettoxford.slack.com/archives/C069SADHP1Q/p1716892511090679) from @alschaffer there appear to be occasions when `rstudio-prefs.json` is empty at the time this script runs. 

In those instances the previous call to `jq` would return a null output and so the working directory preference would not be set.

We don't yet know what these circumstances are.

This PR proposes a way to handle the empty file, we need to carefully test to ensure this setting persists beyond whatever process adds to this file in the circumstances described above.