Use Docker image files in dev container config
See opensafely-core/research-template-docker#32.

In opensafely-core/research-template-docker#57, we moved this
configuration to the Docker image.

We can now update the dev container configuration here to use those
Docker image files.

Note that `postAttachCommand` now becomes a string, instead of an
object; the object format allows for running multiple commands in
parallel:

https://containers.dev/implementors/json_reference/#formatting-string-vs-array-properties

However, for our use case, this is fine for now as we only run
one simple script (that only starts R Studio currently).